Refactored the code and folder structure
