+++ frontend/src/components/custom-tools/output-analyzer/OutputAnalyzer.css
