+++ frontend/src/components/custom-tools/output-analyzer/OutputAnalyzerCard.jsx
